requirements.txt
langchain-groq==0.2.0
langchain-community==0.3.0
faiss-cpu==1.8.0.post1
langchain-text-splitters==0.3.0
streamlit==1.38.0
python-dotenv==1.0.1